refactor: mechanical service moves + atomic backup rename (PR 2e/?, #228)
Three concerns, all part of the Issue #228 strangler-pattern rollout:

1. Move 8 no-db.query service files to their target domain packages
   via `git mv` (blame preserved):
   - simplified_sync.py, real_time_server_commands.py,
     server_properties_generator.py -> app/servers/application/
   - jar_cache_manager.py, java_compatibility.py,
     minecraft_api_service.py, version_manager.py
     -> app/versions/application/
   - encoding_handler.py -> app/files/application/

2. Retain explicit-name re-export shims at the old paths so
   legacy callers (in-app + tests) keep working. Each shim is
   <=13 lines with an explicit __all__ and a single import
   statement (no `from X import *`).

3. Punch-list B (orphan-on-failure backup files): rework
   `BackupService.create_backup` and `BackupService.upload_backup`
   to follow the atomic-rename pattern -- write the tar.gz to
   `.pending/.pending-<uuid>.tar.gz`, commit the DB row, then
   `os.replace()` the temp file into the canonical backups
   directory. If commit fails, the temp file is unlinked and no
   orphan tar ever lands under `backups_directory/`. A new
   regression suite at
   `tests/integration/backups/test_orphan_prevention.py` pins
   the contract for both `create_backup` and `upload_backup`.

Also updates the two PR #2d residual imports flagged by the
plan (`minecraft_server.py: java_compatibility_service` +
`simplified_sync_service`) to point at the new canonical paths;
this is necessary because patches in the unit test files
(`test_real_time_server_commands.py`, `test_java_compatibility_service.py`)
target module-level *internals* (not the re-exported singleton)
and have been updated to the new module paths.

External callsites that go through the re-exported singleton
(e.g. `app.services.jar_cache_manager.jar_cache_manager.copy_jar_to_server`)
keep working unchanged because shim re-exports preserve object
identity.

Refs #228 (PR 2e/?)
